Do not cast to float when calculating double-precision vector Length. Fixes issue [#1542]: "Vector3d.Length casts result to float".

+++ Source/OpenTK/Math/Vector2d.cs
@@ -167,7 +167,7 @@ namespace OpenTK
         {
             get
             {
-                return (float)System.Math.Sqrt(X * X + Y * Y);
+                return System.Math.Sqrt(X * X + Y * Y);
             }
         }
